Add check for the type of MULTITERMINAL symbols.
When there is a rule such as "e ::= ONE|TWO(X)" then ONE and TWO must have
the same type or else the generated code will not work. If the
MULTITERMINAL does not include an alias ("e ::= ONE|TWO") then they are
allowed to have different type, because the generated code will not
reference it.