fix style issues reported by cargo clippy
the rust linter, cargo clippy, complains about a few style issues
that are generated by lemon. this fixes them.

if a variable and the field name are the same, you only have to put it once
like, Parser { ..., extra } instead of Parser { ..., extra: extra }

`return` is unnecessary if it's the last statement in a function
so, `YY_ACTION[i as usize] as i32` instead of `return ....;`